+++ src/components/CardList/CardList.js
@@ -6,7 +6,6 @@ import CardHeader from '@material-ui/core/CardHeader';
 import CardContent from '@material-ui/core/CardContent';
 import CardActions from '@material-ui/core/CardActions';
 import IconButton from '@material-ui/core/IconButton';
-import Collapse from '@material-ui/core/Collapse';
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
 import RefreshIcon from '@material-ui/icons/Refresh';
 import {makeStyles} from '@material-ui/core/styles';
@@ -15,23 +14,21 @@ import List from './List';
 import CardPopoverMenu from '../CardPopoverMenu';
 import LoadingIcon from '../LoadingIcon';
 
-const useStyles = makeStyles(() => {
-  return {
-    cardList: {
-      display: 'inline-block',
-      minWidth: '25rem',
-      width: 'auto',
-    },
-    content: {
-      overflowX: 'hidden',
-      height: '30rem',
-      width: '25rem',
-    },
-    footer: {
-      justifyContent: 'flex-end',
-    },
-  };
-});
+const useStyles = makeStyles(() => ({
+  content: {
+    margin: 0,
+  },
+  loadingIconContainer: {
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center',
+    height: '100%',
+    width: '100%',
+  },
+  footer: {
+    justifyContent: 'flex-end',
+  },
+}));
 
 const useButtonStyles = makeStyles(theme => ({
   base: {
@@ -51,7 +48,7 @@ const useButtonStyles = makeStyles(theme => ({
 function CardList({
   cardStyle,
   cardContentClass,
-  contentHeight,
+  contentHeight = '360px',
   contentWidth,
   contentStyle,
   isExpandable = true,
@@ -102,10 +99,7 @@ function CardList({
   };
 
   return (
-    <Card
-      className={classes.cardList}
-      style={{width: contentWidth, ...cardStyle}}
-    >
+    <Card style={cardStyle}>
       {shouldRenderHeader && (
         <CardHeader
           action={
@@ -133,17 +127,22 @@ function CardList({
           }
         />
       )}
-      <Collapse in={isExpanded} timeout="auto" unmountOnExit>
+      {isExpanded && (
         <CardContent
-          className={`${classes.content} ${cardContentClass}`}
+          className={classnames(classes.content, {
+            [cardContentClass]: Boolean(cardContentClass),
+          })}
           style={{
-            height: contentHeight,
-            width: contentWidth,
+            minHeight: contentHeight,
+            maxHeight: contentHeight,
+            minWidth: contentWidth ?? '300px',
             ...contentStyle,
           }}
         >
           {selectedTab.isLoading ? (
-            <LoadingIcon />
+            <div className={classes.loadingIconContainer}>
+              <LoadingIcon />
+            </div>
           ) : (
             <List
               listItems={selectedTab.listItems}
@@ -155,7 +154,7 @@ function CardList({
             />
           )}
         </CardContent>
-      </Collapse>
+      )}
       {isExpanded && selectedTab.handleRefresh && (
         <CardActions className={classes.footer}>
           <IconButton